Minor refactor and finished ShroomBag
