+++ src/main/java/vazkii/botania/client/gui/bags/SlotMagicPlant.java
@@ -1,5 +1,7 @@
+
 package vazkii.botania.client.gui.bags;
 
+import net.minecraft.block.Block;
 import net.minecraft.inventory.Slot;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
@@ -7,10 +9,10 @@ import vazkii.botania.common.block.ModBlocks;
 
 public abstract class SlotMagicPlant extends Slot {
 
-    InventoryMagicPlant inv;
+    InventoryMagicPlantBag inv;
     int color;
 
-    public SlotMagicPlant(InventoryMagicPlant p_i1824_1_, int p_i1824_2_, int p_i1824_3_, int p_i1824_4_, int color) {
+    public SlotMagicPlant(InventoryMagicPlantBag p_i1824_1_, int p_i1824_2_, int p_i1824_3_, int p_i1824_4_, int color) {
         super(p_i1824_1_, p_i1824_2_, p_i1824_3_, p_i1824_4_);
         this.color = color;
         inv = p_i1824_1_;
@@ -22,6 +24,9 @@ public abstract class SlotMagicPlant extends Slot {
     }
 
     @Override
-    public abstract boolean isItemValid(ItemStack stack);
+    public boolean isItemValid(ItemStack stack) {
+        return stack.getItem() == Item.getItemFromBlock(getValidBlock()) && stack.getItemDamage() == color;
+    }
 
+    protected abstract Block getValidBlock();
 }
