+++ src/main/java/vazkii/botania/common/lib/LibItemNames.java
@@ -126,6 +126,7 @@ public final class LibItemNames {
 	public static final String COSMETIC = "cosmetic";
 	public static final String SWAP_RING = "swapRing";
 	public static final String FLOWER_BAG = "flowerBag";
+	public static final String SHROOM_BAG = "shroomBag";
 	public static final String PHANTOM_INK = "phantomInk";
 	public static final String POOL_MINECART = "poolMinecart";
 	public static final String PINKINATOR = "pinkinator";
