+++ src/main/java/vazkii/botania/common/network/GuiHandler.java
@@ -12,8 +12,10 @@ package vazkii.botania.common.network;
 
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.world.World;
-import vazkii.botania.client.gui.flowerBag.ContainerFlowerBag;
-import vazkii.botania.client.gui.flowerBag.GuiFlowerBag;
+import vazkii.botania.client.gui.bags.flowerBag.ContainerFlowerBag;
+import vazkii.botania.client.gui.bags.flowerBag.GuiFlowerBag;
+import vazkii.botania.client.gui.bags.shroomBag.ContainerShroomBag;
+import vazkii.botania.client.gui.bags.shroomBag.GuiShroomBag;
 import vazkii.botania.client.gui.box.ContainerBaubleBox;
 import vazkii.botania.client.gui.box.GuiBaubleBox;
 import vazkii.botania.client.gui.crafting.ContainerCraftingHalo;
@@ -31,6 +33,8 @@ public class GuiHandler implements IGuiHandler {
 			return new ContainerCraftingHalo(player.inventory, world);
 		case LibGuiIDs.FLOWER_BAG :
 			return new ContainerFlowerBag(player);
+		case LibGuiIDs.SHROOM_BAG:
+			return new ContainerShroomBag(player);
 		case LibGuiIDs.BAUBLE_BOX :
 			return new ContainerBaubleBox(player);
 		}
@@ -48,6 +52,8 @@ public class GuiHandler implements IGuiHandler {
 			return new GuiCraftingHalo(player.inventory, world);
 		case LibGuiIDs.FLOWER_BAG :
 			return new GuiFlowerBag(player);
+		case LibGuiIDs.SHROOM_BAG:
+			return new GuiShroomBag(player);
 		case LibGuiIDs.BAUBLE_BOX :
 			return new GuiBaubleBox(player);
 		}
