+++ src/main/resources/assets/botania/lang/en_US.lang
@@ -1630,6 +1630,7 @@ item.botania:cosmetic31.name=Questgiver Mark
 item.botania:manaString.name=Mana Infused String
 item.botania:swapRing.name=Ring of Correction
 item.botania:flowerBag.name=Flower Pouch
+item.botania:shroomBag.name=Mushroom Pouch
 item.botania:phantomInk.name=Phantom Ink
 item.botania:poolMinecart.name=Minecart with Mana Pool
 item.botania:pinkinator.name=The Pinkinator
